notes added
When to use pointer receiver?
Generally pointer receivers can be used when changes made to the receiver inside the method should be visible to the caller.

Pointers receivers can also be used in places where it's expensive to copy a data structure. Consider a struct which has many fields. Using this struct as a value receiver in a method will need the entire struct to be copied which will be expensive. In this case if a pointer receiver is used, the struct will not be copied and only a pointer to it will be used in the method.

In all other situations value receivers can be used.